requirements.txt
ansible==2.9.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.3.0
PyYAML==5.3.1
requests==2.23.0
scp==0.13.2
six==1.14.0
urllib3==1.25.9
virl2-client==2.0.0b5
